separate variations schema
