+++ functions/routes/variations.js
@@ -14,130 +14,103 @@ const {
   sendSchemaValidationError,
 } = require('../errors/helpers');
 
-const wrap = require('../errors/wrap');
 const { OK } = require('../errors/codes');
+const wrap = require('../errors/wrap');
+const {
+  postSchema,
+  putSchema,
+  variationImage,
+  variationDescription,
+} = require('../models/variations');
 
 // Collection/doc name in Firestore
 const collection = 'variations';
 const docName = 'variation';
 
 /**
- * Schematics for Variation data
- * TODO: move some of the generic ones to the ../schema folder
- */
-const variationId          = Joi.number().integer().min(0);
-const variationName        = Joi.string().min(1);
-const variationDate        = Joi.string().min(4);
-const variationDescription = Joi.array().min(1).items(Joi.string().required());
-const variationImage       = Joi.object({
-                               url:            Joi.string().uri().required(),
-                               caption:        Joi.string().allow("").required(),
-                               componentImage: Joi.bool().required()
-                            });
-const variationImages      = Joi.array().min(1).items(variationImage);
-                            
-
-/**
- * POST /variations schema
- */
-const postSchema = Joi.object({
-  id:          variationId.required(),
-  name:        variationName.required(),
-  date:        variationDate.required(),
-  description: variationDescription.required(),
-  images:      variationImages.required()
-});
-
-// PUT /variations/:id schema
-const putSchema = Joi.object({
-  name:        variationName,
-  date:        variationDate,
-  description: variationDescription,
-  images:      variationImages
-});
-
-/**
  * Variations endpoints
- * 
+ *
  *   POST, GET               /variations
  *         GET, PUT, DELETE  /variations/:id
  *   POST, GET               /variations/:id/images
  *   POST, GET               /variations/:id/description
- * 
+ *
  * TODO: add support for query strings eventually
  */
 
 /**
  * POST /variations
  */
-routes.post('/', wrap(async (req, res, next) => {
+routes.post(
+  '/',
+  wrap(async (req, res, next) => {
+    // Validate request body using schema
+    try {
+      await postSchema.validateAsync(req.body);
+    } catch (e) {
+      return sendSchemaValidationError(res, e);
+    }
 
-  // Validate request body using schema
-  try {
-    await postSchema.validateAsync(req.body);
-  } catch (e) {
-    return sendSchemaValidationError(res, e);
-  }
-  
-  // Try creating a document, and throw error if the doc exists
-  try {
-    await db.collection(collection).doc(`/${req.body.id}/`)
-      .create({
+    // Try creating a document, and throw error if the doc exists
+    try {
+      await db.collection(collection).doc(`/${req.body.id}/`).create({
         id: req.body.id,
         name: req.body.name,
         date: req.body.date,
         description: req.body.description,
         images: req.body.images,
       });
-  } catch (e) {
-    return sendExistingIdError(res, req.body.id, docName);
-  }
-  
-  return res.status(OK).send(req.body);
+    } catch (e) {
+      return sendExistingIdError(res, req.body.id, docName);
+    }
 
-}));
+    return res.status(OK).send(req.body);
+  })
+);
 
 /**
  * GET /variations
  */
-routes.get('/', wrap(async (req, res, next) => {
+routes.get(
+  '/',
+  wrap(async (req, res, next) => {
+    // Query the collection and setup response
+    let query = db.collection(collection);
+    let response = [];
 
-  // Query the collection and setup response
-  let query = db.collection(collection);
-  let response = [];
+    // Get all documents from collection
+    await query.get().then((snapshot) => {
+      let docs = snapshot.docs;
 
-  // Get all documents from collection
-  await query.get().then(snapshot => {
-    let docs = snapshot.docs;
+      for (let variation of docs) {
+        // Insert all data from server doc to response doc
+        const selectedItem = {
+          id: variation.id,
+          name: variation.data().name,
+          date: variation.data().date,
+          description: variation.data().description,
+          images: variation.data().images,
+        };
 
-    for (let variation of docs) {
-      // Insert all data from server doc to response doc
-      const selectedItem = {
-        id: variation.id,
-        name: variation.data().name,
-        date: variation.data().date,
-        description: variation.data().description,
-        images: variation.data().images,
-      };
+        // Put the response doc into the response list
+        response.push(selectedItem);
+      }
 
-      // Put the response doc into the response list
-      response.push(selectedItem);
-    }
-
-    // Send the response once every doc has been put in
-    return res.status(OK).send(response);
-  });
-
-}));
+      // Send the response once every doc has been put in
+      return res.status(OK).send(response);
+    });
+  })
+);
 
 /**
  * GET /variations/:id
  */
-routes.get('/:id', wrap(async (req, res, next) => {
+routes.get(
+  '/:id',
+  wrap(async (req, res, next) => {
+    const document = db.collection(collection).doc(req.params.id);
 
-  const document = db.collection(collection).doc(req.params.id);
-
-  await document.get().then(doc => {
+    await document.get().then((doc) => {
       if (doc.exists) {
         // Fetch and send data if variation of :id is found
         let response = doc.data();
@@ -146,214 +119,217 @@ routes.get('/:id', wrap(async (req, res, next) => {
         // If ID is not found, send error response
         return sendNonexistentIdError(res, req.params.id, docName);
       }
-    }
-  );
-    
-}));
+    });
+  })
+);
 
 /**
  * PUT /variations/:id
  */
-routes.put('/:id', wrap(async (req, res, next) => {
-
-  // Validate request body for correct schema
-  try {
-    await putSchema.validateAsync(req.body);
-  } catch (e) {
-    // If client tries updating id, send immutable attribute error
-    if (e.details[0].context.key === 'id') {
-      return sendImmutableAttributeError(res, 'id');
-    }
-    return sendSchemaValidationError(res, e);
-  }
-
-  const document = db.collection(collection).doc(req.params.id);
-  const docRef = document;
-
-  await document.get().then(doc => {
-    if (!doc.exists) {
-      return sendNonexistentIdError(res, req.params.id, docName);
+routes.put(
+  '/:id',
+  wrap(async (req, res, next) => {
+    // Validate request body for correct schema
+    try {
+      await putSchema.validateAsync(req.body);
+    } catch (e) {
+      // If client tries updating id, send immutable attribute error
+      if (e.details[0].context.key === 'id') {
+        return sendImmutableAttributeError(res, 'id');
+      }
+      return sendSchemaValidationError(res, e);
     }
 
-    docRef.update(req.body);
+    const document = db.collection(collection).doc(req.params.id);
+    const docRef = document;
 
-    // Spread operator to combine old data with updated data
-    // Shared fields are overwritten by rightmost object (updated data)
-    return res.status(OK).send({...doc.data(), ...req.body});
+    await document.get().then((doc) => {
+      if (!doc.exists) {
+        return sendNonexistentIdError(res, req.params.id, docName);
+      }
 
-  });
+      docRef.update(req.body);
 
-  // linting purposes
-  return null;
+      // Spread operator to combine old data with updated data
+      // Shared fields are overwritten by rightmost object (updated data)
+      return res.status(OK).send({ ...doc.data(), ...req.body });
+    });
 
-}));
+    // linting purposes
+    return null;
+  })
+);
 
 /**
  * DELETE /variations/:id
  */
-routes.delete('/:id', wrap(async (req, res, next) => {
-  
-  const document = db.collection(collection).doc(req.params.id);
-  const docRef = document;
+routes.delete(
+  '/:id',
+  wrap(async (req, res, next) => {
+    const document = db.collection(collection).doc(req.params.id);
+    const docRef = document;
 
-  await document.get().then(doc => {
-    if (!doc.exists) {
-      return sendNonexistentIdError(res, req.params.id, docName);
-    }
-    
-    const deletedVariation = doc.data();
-    docRef.delete();
-    return res.status(OK).send(deletedVariation);
-  });
+    await document.get().then((doc) => {
+      if (!doc.exists) {
+        return sendNonexistentIdError(res, req.params.id, docName);
+      }
 
-}));
+      const deletedVariation = doc.data();
+      docRef.delete();
+      return res.status(OK).send(deletedVariation);
+    });
+  })
+);
 
 /**
  * POST /variations/:id/images
  */
-routes.post('/:id/images', wrap(async (req, res, next) => {
+routes.post(
+  '/:id/images',
+  wrap(async (req, res, next) => {
+    // Initialize list of images to add
+    let newImages = [];
 
-  // Initialize list of images to add
-  let newImages = [];
+    // Check whether request body is an array of images,
+    // a single image, or neither
+    try {
+      if (Array.isArray(req.body)) {
+        // Validate all image objects before adding
+        for (const image of req.body) {
+          const validation = variationImage.validate(image);
 
-  // Check whether request body is an array of images, 
-  // a single image, or neither
-  try {
+          // Throw error if image is invalid
+          if (validation.error !== undefined) {
+            throw validation.error;
+          } else if (validation.errors !== undefined) {
+            throw validation.errors;
+          }
 
-    if (Array.isArray(req.body)) {
-      // Validate all image objects before adding
-      for (const image of req.body) {
-        const validation = variationImage.validate(image);
-
-        // Throw error if image is invalid
-        if (validation.error !== undefined) {
-          throw validation.error;
-        } else if (validation.errors !== undefined) {
-          throw validation.errors;
+          newImages.push(image);
         }
-
-        newImages.push(image);
+      } else if (typeof req.body === 'object') {
+        // Validate single image
+        // Note: validateAsync throws an error for you
+        await variationImage.validateAsync(req.body);
+        newImages.push(req.body);
+      } else {
+        // Send error if request body is incorrect type
+        return sendIncorrectTypeError(
+          res,
+          'Request body must be (array of) Variation image object(s)'
+        );
       }
-    } else if (typeof req.body === "object") {
-      // Validate single image
-      // Note: validateAsync throws an error for you
-      await variationImage.validateAsync(req.body);
-      newImages.push(req.body);
-    } else {
-      // Send error if request body is incorrect type
-      return sendIncorrectTypeError(
-        res, 
-        'Request body must be (array of) Variation image object(s)'
-      );
+    } catch (e) {
+      // Schema validation errors end up here
+      return sendSchemaValidationError(res, e);
     }
 
-  } catch (e) {
-    // Schema validation errors end up here
-    return sendSchemaValidationError(res, e);
-  }
-  
-  // Once all images have been validated, add them to the variation with id
-  let images = [];
-  const document = db.collection(collection).doc(req.params.id);
-  const docRef = document;
+    // Once all images have been validated, add them to the variation with id
+    let images = [];
+    const document = db.collection(collection).doc(req.params.id);
+    const docRef = document;
 
-  await document.get().then(doc => {
-    if (!doc.exists) {
-      return sendNonexistentIdError(res, req.params.id, docName);
-    }
+    await document.get().then((doc) => {
+      if (!doc.exists) {
+        return sendNonexistentIdError(res, req.params.id, docName);
+      }
 
-    // Fetch original images and add new images
-    images = doc.data().images;
+      // Fetch original images and add new images
+      images = doc.data().images;
 
-    for (const newImage of newImages) {
-      images.push(newImage);
-    }
+      for (const newImage of newImages) {
+        images.push(newImage);
+      }
 
-    docRef.update({images: images});
-    return res.status(OK).send(images);
+      docRef.update({ images: images });
+      return res.status(OK).send(images);
+    });
 
-  });
-
-  // linting purposes
-  return null;
-
-}));
+    // linting purposes
+    return null;
+  })
+);
 
 /**
  * GET /variations/:id/images
  */
-routes.get('/:id/images', wrap(async (req, res, next) => {
+routes.get(
+  '/:id/images',
+  wrap(async (req, res, next) => {
+    const document = db.collection(collection).doc(req.params.id);
 
-  const document = db.collection(collection).doc(req.params.id);
-  
-  await document.get().then(doc => {
-    if (doc.exists) {
-      return res.status(OK).send(doc.data().images);
-    } else {
-      return sendNonexistentIdError(res, req.params.id, docName);
-    }
-  });
-
-}));
+    await document.get().then((doc) => {
+      if (doc.exists) {
+        return res.status(OK).send(doc.data().images);
+      } else {
+        return sendNonexistentIdError(res, req.params.id, docName);
+      }
+    });
+  })
+);
 
 /**
  * POST /variations/:id/description
  */
-routes.post('/:id/description', wrap(async (req, res, next) => {
+routes.post(
+  '/:id/description',
+  wrap(async (req, res, next) => {
+    let newParagraphs = [];
 
-  let newParagraphs = [];
-  
-  // Check that the body is either string or array of strings, and
-  // add body contents to 'additions' if valid
-  if (typeof req.body === 'string' || req.body instanceof String) {
-    newParagraphs.push(req.body);
-  } else if (Array.isArray(req.body)) {
-    await variationDescription.validateAsync(req.body);
-    newParagraphs = req.body;
-  } else {
-    return sendIncorrectTypeError(res, 'Body must be string or array of strings');
-  }
-
-  const document = db.collection(collection).doc(req.params.id);
-  const docRef = document;
-
-  await document.get().then(doc => {
-    if (!doc.exists) {
-      return sendNonexistentIdError(res, req.params.id, docName);
+    // Check that the body is either string or array of strings, and
+    // add body contents to 'additions' if valid
+    if (typeof req.body === 'string' || req.body instanceof String) {
+      newParagraphs.push(req.body);
+    } else if (Array.isArray(req.body)) {
+      await variationDescription.validateAsync(req.body);
+      newParagraphs = req.body;
+    } else {
+      return sendIncorrectTypeError(
+        res,
+        'Body must be string or array of strings'
+      );
     }
 
-    // Get the current description and add the additions
-    let desc = doc.data().description;
-    for (let paragraph of newParagraphs) {
-      desc.push(paragraph);
-    }
+    const document = db.collection(collection).doc(req.params.id);
+    const docRef = document;
 
-    // Update the description and send response
-    docRef.update({description: desc});
-    return res.status(OK).send(desc);
+    await document.get().then((doc) => {
+      if (!doc.exists) {
+        return sendNonexistentIdError(res, req.params.id, docName);
+      }
 
-  });
+      // Get the current description and add the additions
+      let desc = doc.data().description;
+      for (let paragraph of newParagraphs) {
+        desc.push(paragraph);
+      }
 
-  // linting purposes
-  return null;
-    
-}));
+      // Update the description and send response
+      docRef.update({ description: desc });
+      return res.status(OK).send(desc);
+    });
+
+    // linting purposes
+    return null;
+  })
+);
 
 /**
  * GET /variations/:id/description
  */
-routes.get('/:id/description', wrap(async (req, res, next) => {
+routes.get(
+  '/:id/description',
+  wrap(async (req, res, next) => {
+    const document = db.collection(collection).doc(req.params.id);
 
-  const document = db.collection(collection).doc(req.params.id);
-  
-  await document.get().then(doc => {
-    if (doc.exists) {
-      return res.status(OK).send(doc.data().description);
-    } else {
-      return sendNonexistentIdError(res, req.params.id, docName);
-    }
-  });
-
-}));
+    await document.get().then((doc) => {
+      if (doc.exists) {
+        return res.status(OK).send(doc.data().description);
+      } else {
+        return sendNonexistentIdError(res, req.params.id, docName);
+      }
+    });
+  })
+);
 
 module.exports = routes;
